requirements.txt
annotated-types==0.7.0
anyio
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.61.0
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
litellm==1.61.20
black
blackdoc
isort
autoflake
mcp

lancedb==0.20.0
langchain==0.2.17
langchain_community==0.2.19

# for document loaders
pypdf==5.3.0
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
networkx==3.4.2
python-magic==0.4.27
unstructured==0.16.23

# for document embedding & summarization
# langchain-ollama
langchain-openai==0.1.20

# for using postgres to manage chat history
sqlmodel==0.0.23
asyncpg==0.30.0
SQLAlchemy[asyncpg]==2.0.40
hkdf==0.0.3
python-jose==3.4.0

pandas==2.2.3
gdown==5.2.0

mem0ai==0.1.70
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
json-schema-to-pydantic==0.2.4

docker==7.1.0

# for using opentelemetry tracing
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc

pytest==8.3.5
pytest-asyncio==0.26.0

# for hirag
hnswlib==0.7.0
xxhash==3.5.0